mod: db: drop custom db context wrapper
We want medical profile to be available even if device is credential
encrypted to be able to show the medical profile on lock screen, even
if the device is not unlocked by the user.

The current solution was to split the databases to be either in
/data/system_ce/health or /data/system_de/health.
Though this requires a lot of special sepolicy rules and platform
changes, which can also cause other issues, offer security holes etc.

Instead of doing the heavy lifting of CE vs DE by ourselves, use what
the platform already provides:
 - get rid of the special DbContextWrapper handling
 - pass in CE to SQLiteOpenHelper for HealthStore
 - pass in DE to SQLiteOpenHelper for MedicalProfile

(also fix a typo "heathStore" vs "healthStore")

Change-Id: I33f5c1c50bf7a8aba598874975d95d59fe4cd201
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>