mod, sdk: introduce versioning & validation
It may happen that the mod version installed on a device
and sdk version of an app are not matching, add support
for versioned data by validating it before a db operation.

Validation checks may reset invalid values to their defaults
to avoid having bad data that might cause unexpected results
on db upgrades or other situations.

Signed-off-by: Joey <joey@lineageos.org>