sdk: properly implement getAll and deleteAll
GetAll as it was implemented was not functionally
correct as it didn't take into account the access
control mechanism. Instead let's manually add every
metric type so that access control can verify whether
the calling process actually has access to the
queried metric.

DeleteAll is implemented as a batch operation
so it also respects the access control rules, but
it's made available only for tests

Signed-off-by: Joey <joey@lineageos.org>